+++ src/main/java/com/botmaker/sdk/internal/plugin/editors/ColorEditors.java
@@ -1,7 +1,7 @@
 package com.botmaker.sdk.internal.plugin.editors;
 
 import com.botmaker.plugin.api.StudioServices;
-import com.botmaker.plugin.api.ValueContext;
+import com.botmaker.plugin.api.slot.ValueContext;
 import com.botmaker.plugin.toolkit.Pills;
 import com.botmaker.plugin.toolkit.Slots;
 import com.botmaker.sdk.authoring.WireText;